Fixed weird cuboid offset positions
This basically cleans up the translating, rotating and scaling to make much more sense when we're working with blocks as units.

Before we had a scale of 25%, so a block would be 0.5 wide in terms of GL space. This has been fixed as we no longer do any additional translating or scaling.

Additionally the camera has been set to a much nicer starting position in which we see that the viewport can indeed be rotated (before we had only a 45° downwards rotation).